requirements.txt
discord.py~=2.0.0
motor~=2.4.0
requests~=2.28.1
dnspython~=2.1.0
tabulate~=0.8.9
redis~=3.5.3
aiosqlite~=0.17.0